+++ sdk/compute/arm-compute/src/models/index.ts
@@ -3712,7 +3712,7 @@ export interface GalleryTargetExtendedLocation {
   /** The number of replicas of the Image Version to be created per extended location. This property is updatable. */
   extendedLocationReplicaCount?: number;
   /** Specifies the storage account type to be used to store the image. This property is not updatable. */
-  storageAccountType?: StorageAccountType;
+  storageAccountType?: EdgeZoneStorageAccountType;
   /** Optional. Allows users to provide customer managed keys for encrypting the OS and data disks in the gallery artifact. */
   encryption?: EncryptionImages;
 }
@@ -3980,6 +3980,16 @@ export interface PirCommunityGalleryResource {
   readonly type?: string;
   /** The unique id of this community gallery. */
   uniqueId?: string;
+}
+
+/** This is the community gallery image definition identifier. */
+export interface CommunityGalleryImageIdentifier {
+  /** The name of the gallery image definition publisher. */
+  publisher?: string;
+  /** The name of the gallery image definition offer. */
+  offer?: string;
+  /** The name of the gallery image definition SKU. */
+  sku?: string;
 }
 
 /** The List Community Gallery Images operation response. */
@@ -6513,8 +6523,8 @@ export interface CommunityGalleryImage extends PirCommunityGalleryResource {
   osState?: OperatingSystemStateTypes;
   /** The end of life date of the gallery image definition. This property can be used for decommissioning purposes. This property is updatable. */
   endOfLifeDate?: Date;
-  /** This is the gallery image definition identifier. */
-  identifier?: GalleryImageIdentifier;
+  /** This is the community gallery image definition identifier. */
+  identifier?: CommunityGalleryImageIdentifier;
   /** The properties describe the recommended machine configuration for this Image Definition. These properties are updatable. */
   recommended?: RecommendedMachineConfiguration;
   /** Describes the disallowed disk types. */
@@ -8818,9 +8828,7 @@ export enum KnownStorageAccountType {
   /** StandardZRS */
   StandardZRS = "Standard_ZRS",
   /** PremiumLRS */
-  PremiumLRS = "Premium_LRS",
-  /** StandardSSDLRS */
-  StandardSSDLRS = "StandardSSD_LRS"
+  PremiumLRS = "Premium_LRS"
 }
 
 /**
@@ -8830,8 +8838,7 @@ export enum KnownStorageAccountType {
  * ### Known values supported by the service
  * **Standard_LRS** \
  * **Standard_ZRS** \
- * **Premium_LRS** \
- * **StandardSSD_LRS**
+ * **Premium_LRS**
  */
 export type StorageAccountType = string;
 
@@ -8891,6 +8898,30 @@ export enum KnownGalleryExtendedLocationType {
  * **Unknown**
  */
 export type GalleryExtendedLocationType = string;
+
+/** Known values of {@link EdgeZoneStorageAccountType} that the service accepts. */
+export enum KnownEdgeZoneStorageAccountType {
+  /** StandardLRS */
+  StandardLRS = "Standard_LRS",
+  /** StandardZRS */
+  StandardZRS = "Standard_ZRS",
+  /** StandardSSDLRS */
+  StandardSSDLRS = "StandardSSD_LRS",
+  /** PremiumLRS */
+  PremiumLRS = "Premium_LRS"
+}
+
+/**
+ * Defines values for EdgeZoneStorageAccountType. \
+ * {@link KnownEdgeZoneStorageAccountType} can be used interchangeably with EdgeZoneStorageAccountType,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Standard_LRS** \
+ * **Standard_ZRS** \
+ * **StandardSSD_LRS** \
+ * **Premium_LRS**
+ */
+export type EdgeZoneStorageAccountType = string;
 
 /** Known values of {@link PolicyViolationCategory} that the service accepts. */
 export enum KnownPolicyViolationCategory {