+++ sdk/compute/arm-compute/src/models/mappers.ts
@@ -10404,6 +10404,33 @@ export const PirCommunityGalleryResource: coreClient.CompositeMapper = {
   }
 };
 
+export const CommunityGalleryImageIdentifier: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "CommunityGalleryImageIdentifier",
+    modelProperties: {
+      publisher: {
+        serializedName: "publisher",
+        type: {
+          name: "String"
+        }
+      },
+      offer: {
+        serializedName: "offer",
+        type: {
+          name: "String"
+        }
+      },
+      sku: {
+        serializedName: "sku",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const CommunityGalleryImageList: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -16706,7 +16733,7 @@ export const CommunityGalleryImage: coreClient.CompositeMapper = {
         serializedName: "properties.identifier",
         type: {
           name: "Composite",
-          className: "GalleryImageIdentifier"
+          className: "CommunityGalleryImageIdentifier"
         }
       },
       recommended: {